Switch to using Jekyll to generate site
I’m really just using it as a glorified Markdown pre-processor at the
moment. Many of my existing notes are in Markdown, and it’s a lot more
pleasant than writing HTML.

It's generated using jekyll new --blank, since I want to write all the
layouts and styles myself.

Along the same lines, I replaced the generated
site/_layouts/default.html with the structure from my index.html, and
replaced the generated site/_sass/main.scss with the content of my
styles.css.

I’m using the `site` subdirectory as Jekyll’s root, to make it easy to
separate supporting files (e.g. node_modules) from the actual content,
so that we don’t need to maintain a huge list of things for Jekyll to
ignore. GitHub Pages only lets you serve from the repo’s root or the
/docs directory. Tried creating a symlink but that didn’t work. So, will
be switching to hosting on Netlify. Alternative would be to use a GitHub
Action to push to a gh-pages branch, like the dxw Playbook does [1].

Using the Scripts to Rule Them All pattern.

[1] https://github.com/dxw/playbook/blob/59e54c9905266712fa8e647cad5350762c032da8/.github/workflows/deploy.yml#L52