Tell mobile browsers to use device width for viewport
Based on my quick read of the MDN documentation [1], from where I took
this incantation: when mobile browsers first appeared, back before
anyone was thinking about designing responsive layouts that worked at
small widths, they had to be able to handle big wide pages designed for
desktop. So they'd pretend to have a desktop size viewport and then
scale it down, the intention being that the user would be able to see
the whole page and then zoom in.

But this looks bad on simple layouts like the one we use here. It means
that you get long, illegible lines instead of short, readable ones. This
meta tag tells mobile browsers not to use a fake width and not to do any
scaling.

[1] https://developer.mozilla.org/en-US/docs/Web/HTML/Viewport_meta_tag#viewport_basics